% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_analysis.R
\name{correlation_analysis}
\alias{correlation_analysis}
\title{correlation_analysis}
\usage{
correlation_analysis(
  x,
  result,
  data.info = c("sample", "subject", "time", "class"),
  title = "VIPs_corr",
  groups = list(),
  T0 = F,
  time0 = NULL,
  FC = F,
  scaling = T
)
}
\arguments{
\item{x}{}

\item{result}{}

\item{data.info}{}

\item{title}{}

\item{groups}{}

\item{T0}{}

\item{time0}{}

\item{FC}{}

\item{scaling}{}
}
\description{
performs correlation analysis and corrplots for each of the groups in the dataset
}
\examples{
correlation_analysis(x)
}
