% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_trend.R
\name{group_trend}
\alias{group_trend}
\title{group_trend}
\usage{
group_trend(
  x,
  result = NULL,
  data.info = c("sample", "subject", "time", "class"),
  title = "",
  T0 = F,
  time0 = NULL,
  FC = F,
  scaling = T
)
}
\arguments{
\item{x}{}

\item{result}{}

\item{data.info}{}

\item{title}{}

\item{T0}{}

\item{time0}{}
}
\value{
p
}
\description{
group_trend
}
\examples{
group_trend(x)
}
