% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_plot.R
\name{trend_plot}
\alias{trend_plot}
\title{Title}
\usage{
trend_plot(
  x,
  result = NULL,
  data.info = c("sample", "subject", "time", "class"),
  title = "",
  T0 = F,
  time0 = NULL
)
}
\arguments{
\item{x}{}

\item{result}{}

\item{data.info}{}

\item{title}{}

\item{T0}{}

\item{time0}{}
}
\description{
Title
}
