% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate}
\alias{validate}
\title{Title}
\usage{
validate(
  x,
  data.info = c("sample", "subject", "time", "class"),
  T0,
  MCV = F,
  controls,
  time0,
  comp = 3,
  ntest = 20,
  LOSOXV = F,
  permutime = F,
  permuclass = F
)
}
\arguments{
\item{x}{}

\item{data.info}{}

\item{T0}{}

\item{MCV}{}

\item{controls}{}

\item{time0}{}

\item{comp}{}

\item{ntest}{}

\item{LOSOXV}{}

\item{permutime}{}

\item{permuclass}{}
}
\value{

}
\description{
Title
}
\examples{
validate(x)
}
